Oh boy, big chungus update this one. Focused on maping connected rooms.
So focus for this commit is the maps ability to draw vertical rooms that
are connected by doors. However, it requiered some refactoring of both
the Position struct as well as the Door class to make for an easier
(but still very unefficient) search between such items.

Future note to self: get better at working with pointers, references and
object permanance. The current solution works, but is most likely
copying things way to much. Must fix later for code efficieny.